+++ src/main/java/com/app/yoribogo/domain/CategoryAVO.java
@@ -1,4 +1,4 @@
-package com.app.yoribogo.domain.vo;
+package com.app.yoribogo.domain;
 
 import lombok.Data;
 import org.springframework.stereotype.Component;
@@ -7,7 +7,7 @@ import java.io.Serializable;
 
 @Component
 @Data
-public class TagVO implements Serializable {
+public class CategoryAVO implements Serializable {
     private Long id;
-    private String tagName;
+    private String CategoryAName;
 }
