+++ src/main/java/com/app/yoribogo/domain/LikeVO.java
@@ -1,4 +1,4 @@
-package com.app.yoribogo.domain.vo;
+package com.app.yoribogo.domain;
 
 import lombok.Data;
 import org.springframework.stereotype.Component;
@@ -7,7 +7,8 @@ import java.io.Serializable;
 
 @Component
 @Data
-public class PurchaseVO implements Serializable {
+public class LikeVO implements Serializable {
     private Long id;
     private Long memberId;
+    private Long postId;
 }
