+++ src/main/java/com/app/yoribogo/domain/PurchaseVO.java
@@ -1,4 +1,4 @@
-package com.app.yoribogo.domain.vo;
+package com.app.yoribogo.domain;
 
 import lombok.Data;
 import org.springframework.stereotype.Component;
@@ -7,9 +7,7 @@ import java.io.Serializable;
 
 @Component
 @Data
-public class TagsVO implements Serializable {
+public class PurchaseVO implements Serializable {
     private Long id;
-    private Long tagId;
-    private Long postId;
-
+    private Long memberId;
 }
