+++ src/main/java/com/app/yoribogo/domain/TagVO.java
@@ -1,0 +1,13 @@
+package com.app.yoribogo.domain;
+
+import lombok.Data;
+import org.springframework.stereotype.Component;
+
+import java.io.Serializable;
+
+@Component
+@Data
+public class TagVO implements Serializable {
+    private Long id;
+    private String tagName;
+}
