+++ src/main/java/com/app/yoribogo/domain/TagsVO.java
@@ -1,4 +1,4 @@
-package com.app.yoribogo.domain.vo;
+package com.app.yoribogo.domain;
 
 import lombok.Data;
 import org.springframework.stereotype.Component;
@@ -7,7 +7,9 @@ import java.io.Serializable;
 
 @Component
 @Data
-public class CategoryAVO implements Serializable {
+public class TagsVO implements Serializable {
     private Long id;
-    private String CategoryAName;
+    private Long tagId;
+    private Long postId;
+
 }
