--- src/main/java/com/app/yoribogo/domain/vo/FavoriteRecipeVO.java
@@ -1,14 +1,0 @@
-package com.app.yoribogo.domain.vo;
-
-import lombok.Data;
-import org.springframework.stereotype.Component;
-
-import java.io.Serializable;
-
-@Component
-@Data
-public class FavoriteRecipeVO implements Serializable {
-    private Long id;
-    private Long memberId;
-    private Long recipeId;
-}
