+++ src/main/resources/config/config.xml
@@ -2,21 +2,21 @@
 <!DOCTYPE configuration PUBLIC "-//mybatis.org//DTD Config 3.0/EN" "http://mybatis.org/dtd/mybatis-3-config.dtd">
 <configuration>
     <typeAliases>
-        <typeAlias type="com.app.yoribogo.domain.vo.CategoryAVO" alias="categoryAVO" />
-        <typeAlias type="com.app.yoribogo.domain.vo.CategoryBVO" alias="categoryBVO" />
-        <typeAlias type="com.app.yoribogo.domain.vo.CategoryCVO" alias="categoryCVO" />
-        <typeAlias type="com.app.yoribogo.domain.vo.CommentVO" alias="commentVO" />
-        <typeAlias type="com.app.yoribogo.domain.vo.FavoriteRecipeVO" alias="favoriteRecipeVO" />
-        <typeAlias type="com.app.yoribogo.domain.vo.LikeVO" alias="likeVO" />
-        <typeAlias type="com.app.yoribogo.domain.vo.MemberVO" alias="memberVO" />
-        <typeAlias type="com.app.yoribogo.domain.vo.MessageVO" alias="messageVO" />
-        <typeAlias type="com.app.yoribogo.domain.vo.PostFileVO" alias="postFileVO" />
-        <typeAlias type="com.app.yoribogo.domain.vo.PostVO" alias="postVO" />
-        <typeAlias type="com.app.yoribogo.domain.vo.ProfileVO" alias="profileVO" />
-        <typeAlias type="com.app.yoribogo.domain.vo.PurchasedRecipeVO" alias="purchasedRecipeVO" />
-        <typeAlias type="com.app.yoribogo.domain.vo.PurchaseVO" alias="purchaseVO" />
-        <typeAlias type="com.app.yoribogo.domain.vo.RecipeVO" alias="recipeVO" />
-        <typeAlias type="com.app.yoribogo.domain.vo.TagVO" alias="tagVO" />
-        <typeAlias type="com.app.yoribogo.domain.vo.TagsVO" alias="tagsVO" />
+        <typeAlias type="com.app.yoribogo.domain.CategoryAVO" alias="categoryAVO" />
+        <typeAlias type="com.app.yoribogo.domain.CategoryBVO" alias="categoryBVO" />
+        <typeAlias type="com.app.yoribogo.domain.CategoryCVO" alias="categoryCVO" />
+        <typeAlias type="com.app.yoribogo.domain.CommentVO" alias="commentVO" />
+        <typeAlias type="com.app.yoribogo.domain.FavoriteRecipeVO" alias="favoriteRecipeVO" />
+        <typeAlias type="com.app.yoribogo.domain.LikeVO" alias="likeVO" />
+        <typeAlias type="com.app.yoribogo.domain.MemberVO" alias="memberVO" />
+        <typeAlias type="com.app.yoribogo.domain.MessageVO" alias="messageVO" />
+        <typeAlias type="com.app.yoribogo.domain.PostFileVO" alias="postFileVO" />
+        <typeAlias type="com.app.yoribogo.domain.PostVO" alias="postVO" />
+        <typeAlias type="com.app.yoribogo.domain.ProfileVO" alias="profileVO" />
+        <typeAlias type="com.app.yoribogo.domain.PurchasedRecipeVO" alias="purchasedRecipeVO" />
+        <typeAlias type="com.app.yoribogo.domain.PurchaseVO" alias="purchaseVO" />
+        <typeAlias type="com.app.yoribogo.domain.RecipeVO" alias="recipeVO" />
+        <typeAlias type="com.app.yoribogo.domain.TagVO" alias="tagVO" />
+        <typeAlias type="com.app.yoribogo.domain.TagsVO" alias="tagsVO" />
     </typeAliases>
 </configuration>
